requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.9.2
cachetools==5.3.1; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_version >= '3.7'
click==8.1.3
email-validator==2.0.0.post2
flask==2.2.2
flask-cors==3.0.10
flask-login==0.6.2
flask-migrate==4.0.2
flask-sqlalchemy==3.0.2
flask-wtf==1.1.1
google-auth-oauthlib==1.0.0
google-auth==2.22.0
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.2.4
markupsafe==2.1.2
oauthlib==3.2.2; python_version >= '3.6'
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.31.0; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
setuptools==68.0.0; python_version >= '3.7'
six==1.16.0
sqlalchemy==1.4.46
urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.2.2
wtforms==3.0.1
zipp==3.16.2; python_version >= '3.8'
